Added a mouse sensitivity slider to the options menu (under "Input"), implemented raw mouse input and changed the camera lock key from left alt to tab, which is more convenient.

+++ Source/Input.hpp
@@ -26,7 +26,7 @@ public:
 	bool leftMouseButtonPressed;
 	bool rightMouseButtonPressed;
 
-	glm::vec2 mouseDelta;
+	glm::ivec2 mouseDelta;
 
 	Input(const std::shared_ptr<Window> window);
